add forking and deleting of testnets in beforeAll and afterAll
