+++ package.json
@@ -16,9 +16,8 @@
     "test": "jest --config=jest.config.js --runInBand --watch",
     "test:ci": "jest --runInBand --ci --coverage --detectOpenHandles --forceExit",
     "e2e": "APP_ENV=test start-server-and-test dev http://localhost:3000 cypress",
-    "e2e:playwright": "yarn vnet:delete:all && yarn vnet:fork && playwright test --trace on",
-    "playwright:no-vnet": "playwright test --trace on",
-    "playwright:ui": "yarn vnet:delete:all && yarn vnet:fork && playwright test --ui",
+    "e2e:playwright": "playwright test --trace on",
+    "playwright:ui": "playwright test --ui",
     "e2e:headless": "APP_ENV=test npm run fund && start-server-and-test dev http://localhost:3000 cypress:headless",
     "cypress": "BABEL_ENV=cypress-test cypress open",
     "cypress:headless": "BABEL_ENV=cypress-test cypress run",
