+++ playwright/example.spec.ts
@@ -1,4 +1,5 @@
 import { test, expect } from '@playwright/test';
+import './forkVnet';
 
 test('first test', async ({ page }) => {
   await page.goto('/');
